FST: Add the RaidMetaLayout to deal with general operations on RAID-like files
Conflicts:

	fst/CMakeLists.txt
	fst/FmdSqlite.cc
	fst/layout/RaidDpLayout.cc
	fst/layout/ReedSLayout.cc

+++ fst/layout/LocalFileIo.cc
@@ -96,7 +96,10 @@ LocalFileIo::Write( XrdSfsFileOffset offset,
                     char*            buffer,
                     XrdSfsXferSize   length )
 {
-  eos_debug( "offset = %llu, length = %lu", offset, length );
+  eos_debug( "offset = %lli, length = %lli",
+             static_cast<int64_t>( offset ),
+             static_cast<int64_t>( length ) );
+
   return mLogicalFile->writeofs( offset, buffer, length );
 }
